+++ config/dev_webpack.config.js
@@ -41,7 +41,7 @@ module.exports = {
       },
       {
         test: /\.jsx$/,
-        loader: 'react-hot!babel'
+        loader: 'babel'
       },
       {
         test: /\.json$/,